+++ azure-iot-hub/tests/test_iothub_registry_manager.py
@@ -5,62 +5,387 @@
 # --------------------------------------------------------------------------
 
 import pytest
-from .device_create_tests import (
-    CreateDeviceWithSymmetricKeyTests,
-    CreateDeviceWithX509Tests,
-    CreateDeviceWithCATests,
-)
-from .device_update_tests import (
-    UpdateDeviceWithSymmetricKeyTests,
-    UpdateDeviceWithX509Tests,
-    UpdateDeviceWithCATests,
-)
-
-from .module_create_tests import (
-    CreateModuleWithSymmetricKeyTests,
-    CreateModuleWithX509Tests,
-    CreateModuleWithCATests,
-)
-from .module_update_tests import (
-    UpdateModuleWithSymmetricKeyTests,
-    UpdateModuleWithX509Tests,
-    UpdateModuleWithCATests,
+from azure.iot.hub.protocol.models import AuthenticationMechanism
+from .common_fixtures import (
+    fake_device_id,
+    fake_module_id,
+    fake_managed_by,
+    fake_status,
+    fake_etag,
+    fake_primary_key,
+    fake_secondary_key,
+    fake_primary_thumbprint,
+    fake_secondary_thumbprint,
 )
 
 
-from .common_fixtures import fake_device_id, fake_etag, fake_module_id
+@pytest.fixture(scope="function")
+def mock_device_constructor(mocker):
+    return mocker.patch("azure.iot.hub.iothub_registry_manager.Device")
 
 
-class DeviceMock(object):
-    @pytest.fixture(scope="function")
-    def mock_device_constructor(self, mocker):
-        return mocker.patch("azure.iot.hub.iothub_registry_manager.Device")
+@pytest.fixture(scope="function")
+def mock_module_constructor(mocker):
+    return mocker.patch("azure.iot.hub.iothub_registry_manager.Module")
 
 
-class ModuleMock(object):
-    @pytest.fixture(scope="function")
-    def mock_module_constructor(self, mocker):
-        return mocker.patch("azure.iot.hub.iothub_registry_manager.Module")
+@pytest.mark.describe("IoTHubRegistryManager - .create_device_with_sas()")
+class TestCreateDeviceWithSymmetricKey(object):
+
+    testdata = [(fake_primary_key, None), (None, fake_secondary_key)]
+
+    @pytest.mark.it("initializes device with device id, status and sas auth")
+    @pytest.mark.parametrize(
+        "primary_key, secondary_key", testdata, ids=["Primary Key", "Secondary Key"]
+    )
+    def test_initializes_device_with_kwargs_for_sas(
+        self,
+        mock_service_operations,
+        iothub_registry_manager,
+        mock_device_constructor,
+        primary_key,
+        secondary_key,
+    ):
+        iothub_registry_manager.create_device_with_sas(
+            device_id=fake_device_id,
+            status=fake_status,
+            primary_key=primary_key,
+            secondary_key=secondary_key,
+        )
+
+        assert mock_device_constructor.call_count == 1
+
+        assert mock_device_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_device_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_device_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_device_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "sas"
+        assert auth_mechanism.x509_thumbprint is None
+        sym_key = auth_mechanism.symmetric_key
+        assert sym_key.primary_key == primary_key
+        assert sym_key.secondary_key == secondary_key
+
+    @pytest.mark.it(
+        "calls method from service operations with device id and previously constructed device"
+    )
+    @pytest.mark.parametrize(
+        "primary_key, secondary_key", testdata, ids=["Primary Key", "Secondary Key"]
+    )
+    def test_calls_create_or_update_device_for_sas(
+        self,
+        mock_device_constructor,
+        mock_service_operations,
+        iothub_registry_manager,
+        primary_key,
+        secondary_key,
+    ):
+        iothub_registry_manager.create_device_with_sas(
+            device_id=fake_device_id,
+            status=fake_status,
+            primary_key=primary_key,
+            secondary_key=secondary_key,
+        )
+
+        assert mock_service_operations.create_or_update_device.call_count == 1
+        assert mock_service_operations.create_or_update_device.call_args[0][0] == fake_device_id
+        assert (
+            mock_service_operations.create_or_update_device.call_args[0][1]
+            == mock_device_constructor.return_value
+        )
 
 
-@pytest.mark.describe("IotHub Registry Manager Device Create Tests")
-class TestIotHubRegistryManagerDeviceCreate(
-    DeviceMock,
-    CreateDeviceWithSymmetricKeyTests,
-    CreateDeviceWithX509Tests,
-    CreateDeviceWithCATests,
-):
-    pass
+@pytest.mark.describe("IoTHubRegistryManager - .create_device_with_x509()")
+class TestCreateDeviceWithX509(object):
+
+    testdata = [(fake_primary_thumbprint, None), (None, fake_secondary_thumbprint)]
+
+    @pytest.mark.it("initializes device with device id, status and X509 auth")
+    @pytest.mark.parametrize(
+        "primary_thumbprint, secondary_thumbprint",
+        testdata,
+        ids=["Primary Thumbprint", "Secondary Thumbprint"],
+    )
+    def test_initializes_device_with_kwargs_for_x509(
+        self,
+        mock_service_operations,
+        iothub_registry_manager,
+        mock_device_constructor,
+        primary_thumbprint,
+        secondary_thumbprint,
+    ):
+        iothub_registry_manager.create_device_with_x509(
+            device_id=fake_device_id,
+            status=fake_status,
+            primary_thumbprint=primary_thumbprint,
+            secondary_thumbprint=secondary_thumbprint,
+        )
+
+        assert mock_device_constructor.call_count == 1
+        assert mock_device_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_device_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_device_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_device_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "selfSigned"
+        assert auth_mechanism.symmetric_key is None
+        x509_thumbprint = auth_mechanism.x509_thumbprint
+        assert x509_thumbprint.primary_thumbprint == primary_thumbprint
+        assert x509_thumbprint.secondary_thumbprint == secondary_thumbprint
+
+    @pytest.mark.it(
+        "calls method from service operations with device id and previously constructed device"
+    )
+    @pytest.mark.parametrize(
+        "primary_thumbprint, secondary_thumbprint",
+        testdata,
+        ids=["Primary Thumbprint", "Secondary Thumbprint"],
+    )
+    def test_calls_create_or_update_device_for_x509(
+        self,
+        mock_device_constructor,
+        mock_service_operations,
+        iothub_registry_manager,
+        primary_thumbprint,
+        secondary_thumbprint,
+    ):
+        iothub_registry_manager.create_device_with_x509(
+            device_id=fake_device_id,
+            status=fake_status,
+            primary_thumbprint=primary_thumbprint,
+            secondary_thumbprint=secondary_thumbprint,
+        )
+
+        assert mock_service_operations.create_or_update_device.call_count == 1
+        assert mock_service_operations.create_or_update_device.call_args[0][0] == fake_device_id
+        assert (
+            mock_service_operations.create_or_update_device.call_args[0][1]
+            == mock_device_constructor.return_value
+        )
 
 
-@pytest.mark.describe("IotHub Registry Manager Device Update Tests")
-class TestIotHubRegistryManagerDeviceUpdate(
-    DeviceMock,
-    UpdateDeviceWithSymmetricKeyTests,
-    UpdateDeviceWithX509Tests,
-    UpdateDeviceWithCATests,
-):
-    pass
+@pytest.mark.describe("IoTHubRegistryManager - .create_device_with_certificate_authority()")
+class TestCreateDeviceWithCA(object):
+    @pytest.mark.it("initializes device with device id, status and ca auth")
+    def test_initializes_device_with_kwargs_for_certificate_authority(
+        self, mock_device_constructor, mock_service_operations, iothub_registry_manager
+    ):
+        iothub_registry_manager.create_device_with_certificate_authority(
+            device_id=fake_device_id, status=fake_status
+        )
+
+        assert mock_device_constructor.call_count == 1
+        assert mock_device_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_device_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_device_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_device_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "certificateAuthority"
+        assert auth_mechanism.x509_thumbprint is None
+        assert auth_mechanism.symmetric_key is None
+
+    @pytest.mark.it(
+        "calls method from service operations with device id and previously constructed device"
+    )
+    def test_calls_create_or_update_device_for_certificate_authority(
+        self, mock_device_constructor, mock_service_operations, iothub_registry_manager
+    ):
+        iothub_registry_manager.create_device_with_certificate_authority(
+            device_id=fake_device_id, status=fake_status
+        )
+
+        assert mock_service_operations.create_or_update_device.call_count == 1
+        assert mock_service_operations.create_or_update_device.call_args[0][0] == fake_device_id
+        assert (
+            mock_service_operations.create_or_update_device.call_args[0][1]
+            == mock_device_constructor.return_value
+        )
+
+
+@pytest.mark.describe("IoTHubRegistryManager - .update_device_with_sas()")
+class TestUpdateDeviceWithSymmetricKey(object):
+
+    testdata = [(fake_primary_key, None), (None, fake_secondary_key)]
+
+    @pytest.mark.it("initializes device with device id, status, etag and sas auth")
+    @pytest.mark.parametrize(
+        "primary_key, secondary_key", testdata, ids=["Primary Key", "Secondary Key"]
+    )
+    def test_initializes_device_with_kwargs_for_sas(
+        self,
+        mock_service_operations,
+        iothub_registry_manager,
+        mock_device_constructor,
+        primary_key,
+        secondary_key,
+    ):
+        iothub_registry_manager.update_device_with_sas(
+            device_id=fake_device_id,
+            status=fake_status,
+            etag=fake_etag,
+            primary_key=primary_key,
+            secondary_key=secondary_key,
+        )
+
+        assert mock_device_constructor.call_count == 1
+
+        assert mock_device_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_device_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_device_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_device_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "sas"
+        assert auth_mechanism.x509_thumbprint is None
+        sym_key = auth_mechanism.symmetric_key
+        assert sym_key.primary_key == primary_key
+        assert sym_key.secondary_key == secondary_key
+        assert mock_device_constructor.call_args[1]["etag"] == fake_etag
+
+    @pytest.mark.it(
+        "calls method from service operations with device id and previously constructed device"
+    )
+    @pytest.mark.parametrize(
+        "primary_key, secondary_key", testdata, ids=["Primary Key", "Secondary Key"]
+    )
+    def test_calls_create_or_update_device_for_sas(
+        self,
+        mock_device_constructor,
+        mock_service_operations,
+        iothub_registry_manager,
+        primary_key,
+        secondary_key,
+    ):
+        iothub_registry_manager.update_device_with_sas(
+            device_id=fake_device_id,
+            status=fake_status,
+            etag=fake_etag,
+            primary_key=primary_key,
+            secondary_key=secondary_key,
+        )
+
+        assert mock_service_operations.create_or_update_device.call_count == 1
+        assert mock_service_operations.create_or_update_device.call_args[0][0] == fake_device_id
+        assert (
+            mock_service_operations.create_or_update_device.call_args[0][1]
+            == mock_device_constructor.return_value
+        )
+
+
+@pytest.mark.describe("IoTHubRegistryManager - .update_device_with_x509()")
+class TestUpdateDeviceWithX509(object):
+
+    testdata = [(fake_primary_thumbprint, None), (None, fake_secondary_thumbprint)]
+
+    @pytest.mark.it("initializes device with device id, status and X509 auth")
+    @pytest.mark.parametrize(
+        "primary_thumbprint, secondary_thumbprint",
+        testdata,
+        ids=["Primary Thumbprint", "Secondary Thumbprint"],
+    )
+    def test_initializes_device_with_kwargs_for_x509(
+        self,
+        mock_service_operations,
+        iothub_registry_manager,
+        mock_device_constructor,
+        primary_thumbprint,
+        secondary_thumbprint,
+    ):
+        iothub_registry_manager.update_device_with_x509(
+            device_id=fake_device_id,
+            status=fake_status,
+            etag=fake_etag,
+            primary_thumbprint=primary_thumbprint,
+            secondary_thumbprint=secondary_thumbprint,
+        )
+
+        assert mock_device_constructor.call_count == 1
+        assert mock_device_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_device_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_device_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_device_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "selfSigned"
+        assert auth_mechanism.symmetric_key is None
+        x509_thumbprint = auth_mechanism.x509_thumbprint
+        assert x509_thumbprint.primary_thumbprint == primary_thumbprint
+        assert x509_thumbprint.secondary_thumbprint == secondary_thumbprint
+        assert mock_device_constructor.call_args[1]["etag"] == fake_etag
+
+    @pytest.mark.it(
+        "calls method from service operations with device id and previously constructed device"
+    )
+    @pytest.mark.parametrize(
+        "primary_thumbprint, secondary_thumbprint",
+        testdata,
+        ids=["Primary Thumbprint", "Secondary Thumbprint"],
+    )
+    def test_calls_create_or_update_device_for_x509(
+        self,
+        mock_device_constructor,
+        mock_service_operations,
+        iothub_registry_manager,
+        primary_thumbprint,
+        secondary_thumbprint,
+    ):
+        iothub_registry_manager.update_device_with_x509(
+            device_id=fake_device_id,
+            status=fake_status,
+            etag=fake_etag,
+            primary_thumbprint=primary_thumbprint,
+            secondary_thumbprint=secondary_thumbprint,
+        )
+
+        assert mock_service_operations.create_or_update_device.call_count == 1
+        assert mock_service_operations.create_or_update_device.call_args[0][0] == fake_device_id
+        assert (
+            mock_service_operations.create_or_update_device.call_args[0][1]
+            == mock_device_constructor.return_value
+        )
+
+
+@pytest.mark.describe("IoTHubRegistryManager - .update_device_with_certificate_authority()")
+class TestUpdateDeviceWithCA(object):
+    @pytest.mark.it("initializes device with device id, status and ca auth")
+    def test_initializes_device_with_kwargs_for_certificate_authority(
+        self, mock_device_constructor, mock_service_operations, iothub_registry_manager
+    ):
+        iothub_registry_manager.update_device_with_certificate_authority(
+            device_id=fake_device_id, status=fake_status, etag=fake_etag
+        )
+
+        assert mock_device_constructor.call_count == 1
+        assert mock_device_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_device_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_device_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_device_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "certificateAuthority"
+        assert auth_mechanism.x509_thumbprint is None
+        assert auth_mechanism.symmetric_key is None
+        assert mock_device_constructor.call_args[1]["etag"] == fake_etag
+
+    @pytest.mark.it(
+        "calls method from service operations with device id and previously constructed device"
+    )
+    def test_calls_create_or_update_device_for_certificate_authority(
+        self, mock_device_constructor, mock_service_operations, iothub_registry_manager
+    ):
+        iothub_registry_manager.update_device_with_certificate_authority(
+            device_id=fake_device_id, status=fake_status, etag=fake_etag
+        )
+
+        assert mock_service_operations.create_or_update_device.call_count == 1
+        assert mock_service_operations.create_or_update_device.call_args[0][0] == fake_device_id
+        assert (
+            mock_service_operations.create_or_update_device.call_args[0][1]
+            == mock_device_constructor.return_value
+        )
 
 
 @pytest.mark.describe("IoTHubRegistryManager -- .get_device()")
@@ -100,24 +425,412 @@ class TestDeleteDevice(object):
         )
 
 
-@pytest.mark.describe("IotHub Registry Manager Module Create Tests")
-class TestIotHubRegistryManagerModuleCreate(
-    ModuleMock,
-    CreateModuleWithSymmetricKeyTests,
-    CreateModuleWithX509Tests,
-    CreateModuleWithCATests,
-):
-    pass
+@pytest.mark.describe("IoTHubRegistryManager - .create_module_with_sas()")
+class TestCreateModuleWithSymmetricKey(object):
+
+    testdata = [(fake_primary_key, None), (None, fake_secondary_key)]
+
+    @pytest.mark.it("initializes module with device id, module id, managed_by status and sas auth")
+    @pytest.mark.parametrize(
+        "primary_key, secondary_key", testdata, ids=["Primary Key", "Secondary Key"]
+    )
+    def test_initializes_device_with_kwargs_for_sas(
+        self,
+        mock_service_operations,
+        iothub_registry_manager,
+        mock_module_constructor,
+        primary_key,
+        secondary_key,
+    ):
+        iothub_registry_manager.create_module_with_sas(
+            device_id=fake_device_id,
+            module_id=fake_module_id,
+            managed_by=fake_managed_by,
+            status=fake_status,
+            primary_key=primary_key,
+            secondary_key=secondary_key,
+        )
+
+        assert mock_module_constructor.call_count == 1
+
+        assert mock_module_constructor.call_args[1]["module_id"] == fake_module_id
+        assert mock_module_constructor.call_args[1]["managed_by"] == fake_managed_by
+        assert mock_module_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_module_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_module_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_module_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "sas"
+        assert auth_mechanism.x509_thumbprint is None
+        sym_key = auth_mechanism.symmetric_key
+        assert sym_key.primary_key == primary_key
+        assert sym_key.secondary_key == secondary_key
+
+    @pytest.mark.it(
+        "calls method from service operations with device id, module id and previously constructed module"
+    )
+    @pytest.mark.parametrize(
+        "primary_key, secondary_key", testdata, ids=["Primary Key", "Secondary Key"]
+    )
+    def test_calls_create_or_update_device_for_sas(
+        self,
+        mock_module_constructor,
+        mock_service_operations,
+        iothub_registry_manager,
+        primary_key,
+        secondary_key,
+    ):
+        iothub_registry_manager.create_module_with_sas(
+            device_id=fake_device_id,
+            status=fake_status,
+            module_id=fake_module_id,
+            managed_by=fake_managed_by,
+            primary_key=primary_key,
+            secondary_key=secondary_key,
+        )
+
+        assert mock_service_operations.create_or_update_module.call_count == 1
+        assert mock_service_operations.create_or_update_module.call_args[0][0] == fake_device_id
+        assert mock_service_operations.create_or_update_module.call_args[0][1] == fake_module_id
+        assert (
+            mock_service_operations.create_or_update_module.call_args[0][2]
+            == mock_module_constructor.return_value
+        )
 
 
-@pytest.mark.describe("IotHub Registry Manager Module Update Tests")
-class TestIotHubRegistryManagerModuleUpdate(
-    ModuleMock,
-    UpdateModuleWithSymmetricKeyTests,
-    UpdateModuleWithX509Tests,
-    UpdateModuleWithCATests,
-):
-    pass
+@pytest.mark.describe("IoTHubRegistryManager - .create_module_with_x509()")
+class TestCreateModuleWithX509(object):
+
+    testdata = [(fake_primary_thumbprint, None), (None, fake_secondary_thumbprint)]
+
+    @pytest.mark.it(
+        "initializes module with device id, module id, managed_by, status and X509 auth"
+    )
+    @pytest.mark.parametrize(
+        "primary_thumbprint, secondary_thumbprint",
+        testdata,
+        ids=["Primary Thumbprint", "Secondary Thumbprint"],
+    )
+    def test_initializes_device_with_kwargs_for_x509(
+        self,
+        mock_service_operations,
+        iothub_registry_manager,
+        mock_module_constructor,
+        primary_thumbprint,
+        secondary_thumbprint,
+    ):
+        iothub_registry_manager.create_module_with_x509(
+            device_id=fake_device_id,
+            status=fake_status,
+            module_id=fake_module_id,
+            managed_by=fake_managed_by,
+            primary_thumbprint=primary_thumbprint,
+            secondary_thumbprint=secondary_thumbprint,
+        )
+
+        assert mock_module_constructor.call_count == 1
+        assert mock_module_constructor.call_args[1]["module_id"] == fake_module_id
+        assert mock_module_constructor.call_args[1]["managed_by"] == fake_managed_by
+        assert mock_module_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_module_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_module_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_module_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "selfSigned"
+        assert auth_mechanism.symmetric_key is None
+        x509_thumbprint = auth_mechanism.x509_thumbprint
+        assert x509_thumbprint.primary_thumbprint == primary_thumbprint
+        assert x509_thumbprint.secondary_thumbprint == secondary_thumbprint
+
+    @pytest.mark.it(
+        "calls method from service operations with device id, module id and previously constructed module"
+    )
+    @pytest.mark.parametrize(
+        "primary_thumbprint, secondary_thumbprint",
+        testdata,
+        ids=["Primary Thumbprint", "Secondary Thumbprint"],
+    )
+    def test_calls_create_or_update_device_for_x509(
+        self,
+        mock_module_constructor,
+        mock_service_operations,
+        iothub_registry_manager,
+        primary_thumbprint,
+        secondary_thumbprint,
+    ):
+        iothub_registry_manager.create_module_with_x509(
+            device_id=fake_device_id,
+            status=fake_status,
+            module_id=fake_module_id,
+            managed_by=fake_managed_by,
+            primary_thumbprint=primary_thumbprint,
+            secondary_thumbprint=secondary_thumbprint,
+        )
+
+        assert mock_service_operations.create_or_update_module.call_count == 1
+        assert mock_service_operations.create_or_update_module.call_args[0][0] == fake_device_id
+        assert mock_service_operations.create_or_update_module.call_args[0][1] == fake_module_id
+        assert (
+            mock_service_operations.create_or_update_module.call_args[0][2]
+            == mock_module_constructor.return_value
+        )
+
+
+@pytest.mark.describe("IoTHubRegistryManager - .create_module_with_certificate_authority()")
+class TestCreateModuleWithCA(object):
+    @pytest.mark.it("initializes module with device id, module id, managed_by, status and ca auth")
+    def test_initializes_device_with_kwargs_for_certificate_authority(
+        self, mock_module_constructor, mock_service_operations, iothub_registry_manager
+    ):
+        iothub_registry_manager.create_module_with_certificate_authority(
+            device_id=fake_device_id,
+            status=fake_status,
+            module_id=fake_module_id,
+            managed_by=fake_managed_by,
+        )
+
+        assert mock_module_constructor.call_count == 1
+        assert mock_module_constructor.call_args[1]["module_id"] == fake_module_id
+        assert mock_module_constructor.call_args[1]["managed_by"] == fake_managed_by
+        assert mock_module_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_module_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_module_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_module_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "certificateAuthority"
+        assert auth_mechanism.x509_thumbprint is None
+        assert auth_mechanism.symmetric_key is None
+
+    @pytest.mark.it(
+        "calls method from service operations with device id, module id and previously constructed module"
+    )
+    def test_calls_create_or_update_device_for_certificate_authority(
+        self, mock_module_constructor, mock_service_operations, iothub_registry_manager
+    ):
+        iothub_registry_manager.create_module_with_certificate_authority(
+            device_id=fake_device_id,
+            status=fake_status,
+            module_id=fake_module_id,
+            managed_by=fake_managed_by,
+        )
+
+        assert mock_service_operations.create_or_update_module.call_count == 1
+        assert mock_service_operations.create_or_update_module.call_args[0][0] == fake_device_id
+        assert mock_service_operations.create_or_update_module.call_args[0][1] == fake_module_id
+        assert (
+            mock_service_operations.create_or_update_module.call_args[0][2]
+            == mock_module_constructor.return_value
+        )
+
+
+@pytest.mark.describe("IoTHubRegistryManager - .update_module_with_sas()")
+class TestUpdateModuleWithSymmetricKey(object):
+
+    testdata = [(fake_primary_key, None), (None, fake_secondary_key)]
+
+    @pytest.mark.it("initializes module with device id, module id, managed_by status and sas auth")
+    @pytest.mark.parametrize(
+        "primary_key, secondary_key", testdata, ids=["Primary Key", "Secondary Key"]
+    )
+    def test_initializes_device_with_kwargs_for_sas(
+        self,
+        mock_service_operations,
+        iothub_registry_manager,
+        mock_module_constructor,
+        primary_key,
+        secondary_key,
+    ):
+        iothub_registry_manager.update_module_with_sas(
+            device_id=fake_device_id,
+            module_id=fake_module_id,
+            managed_by=fake_managed_by,
+            status=fake_status,
+            etag=fake_etag,
+            primary_key=primary_key,
+            secondary_key=secondary_key,
+        )
+
+        assert mock_module_constructor.call_count == 1
+
+        assert mock_module_constructor.call_args[1]["module_id"] == fake_module_id
+        assert mock_module_constructor.call_args[1]["managed_by"] == fake_managed_by
+        assert mock_module_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_module_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_module_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_module_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "sas"
+        assert auth_mechanism.x509_thumbprint is None
+        sym_key = auth_mechanism.symmetric_key
+        assert sym_key.primary_key == primary_key
+        assert sym_key.secondary_key == secondary_key
+
+    @pytest.mark.it(
+        "calls method from service operations with device id, module id and previously constructed module"
+    )
+    @pytest.mark.parametrize(
+        "primary_key, secondary_key", testdata, ids=["Primary Key", "Secondary Key"]
+    )
+    def test_calls_create_or_update_device_for_sas(
+        self,
+        mock_module_constructor,
+        mock_service_operations,
+        iothub_registry_manager,
+        primary_key,
+        secondary_key,
+    ):
+        iothub_registry_manager.update_module_with_sas(
+            device_id=fake_device_id,
+            module_id=fake_module_id,
+            status=fake_status,
+            etag=fake_etag,
+            managed_by=fake_managed_by,
+            primary_key=primary_key,
+            secondary_key=secondary_key,
+        )
+
+        assert mock_service_operations.create_or_update_module.call_count == 1
+        assert mock_service_operations.create_or_update_module.call_args[0][0] == fake_device_id
+        assert mock_service_operations.create_or_update_module.call_args[0][1] == fake_module_id
+        assert (
+            mock_service_operations.create_or_update_module.call_args[0][2]
+            == mock_module_constructor.return_value
+        )
+
+
+@pytest.mark.describe("IoTHubRegistryManager - .update_module_with_x509()")
+class TestUpdateModuleWithX509(object):
+
+    testdata = [(fake_primary_thumbprint, None), (None, fake_secondary_thumbprint)]
+
+    @pytest.mark.it(
+        "initializes module with device id, module id, managed_by, status and X509 auth"
+    )
+    @pytest.mark.parametrize(
+        "primary_thumbprint, secondary_thumbprint",
+        testdata,
+        ids=["Primary Thumbprint", "Secondary Thumbprint"],
+    )
+    def test_initializes_device_with_kwargs_for_x509(
+        self,
+        mock_service_operations,
+        iothub_registry_manager,
+        mock_module_constructor,
+        primary_thumbprint,
+        secondary_thumbprint,
+    ):
+        iothub_registry_manager.update_module_with_x509(
+            device_id=fake_device_id,
+            module_id=fake_module_id,
+            status=fake_status,
+            etag=fake_etag,
+            managed_by=fake_managed_by,
+            primary_thumbprint=primary_thumbprint,
+            secondary_thumbprint=secondary_thumbprint,
+        )
+
+        assert mock_module_constructor.call_count == 1
+        assert mock_module_constructor.call_args[1]["module_id"] == fake_module_id
+        assert mock_module_constructor.call_args[1]["managed_by"] == fake_managed_by
+        assert mock_module_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_module_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_module_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_module_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "selfSigned"
+        assert auth_mechanism.symmetric_key is None
+        x509_thumbprint = auth_mechanism.x509_thumbprint
+        assert x509_thumbprint.primary_thumbprint == primary_thumbprint
+        assert x509_thumbprint.secondary_thumbprint == secondary_thumbprint
+
+    @pytest.mark.it(
+        "calls method from service operations with device id, module id and previously constructed module"
+    )
+    @pytest.mark.parametrize(
+        "primary_thumbprint, secondary_thumbprint",
+        testdata,
+        ids=["Primary Thumbprint", "Secondary Thumbprint"],
+    )
+    def test_calls_create_or_update_device_for_x509(
+        self,
+        mock_module_constructor,
+        mock_service_operations,
+        iothub_registry_manager,
+        primary_thumbprint,
+        secondary_thumbprint,
+    ):
+        iothub_registry_manager.update_module_with_x509(
+            device_id=fake_device_id,
+            module_id=fake_module_id,
+            status=fake_status,
+            etag=fake_etag,
+            managed_by=fake_managed_by,
+            primary_thumbprint=primary_thumbprint,
+            secondary_thumbprint=secondary_thumbprint,
+        )
+
+        assert mock_service_operations.create_or_update_module.call_count == 1
+        assert mock_service_operations.create_or_update_module.call_args[0][0] == fake_device_id
+        assert mock_service_operations.create_or_update_module.call_args[0][1] == fake_module_id
+        assert (
+            mock_service_operations.create_or_update_module.call_args[0][2]
+            == mock_module_constructor.return_value
+        )
+
+
+@pytest.mark.describe("IoTHubRegistryManager - .update_module_with_certificate_authority()")
+class TestUpdateModuleWithCA(object):
+    @pytest.mark.it("initializes module with device id, module id, managed_by, status and ca auth")
+    def test_initializes_device_with_kwargs_for_certificate_authority(
+        self, mock_module_constructor, mock_service_operations, iothub_registry_manager
+    ):
+        iothub_registry_manager.update_module_with_certificate_authority(
+            device_id=fake_device_id,
+            module_id=fake_module_id,
+            etag=fake_etag,
+            status=fake_status,
+            managed_by=fake_managed_by,
+        )
+
+        assert mock_module_constructor.call_count == 1
+        assert mock_module_constructor.call_args[1]["module_id"] == fake_module_id
+        assert mock_module_constructor.call_args[1]["managed_by"] == fake_managed_by
+        assert mock_module_constructor.call_args[1]["device_id"] == fake_device_id
+        assert mock_module_constructor.call_args[1]["status"] == fake_status
+        assert isinstance(
+            mock_module_constructor.call_args[1]["authentication"], AuthenticationMechanism
+        )
+        auth_mechanism = mock_module_constructor.call_args[1]["authentication"]
+        assert auth_mechanism.type == "certificateAuthority"
+        assert auth_mechanism.x509_thumbprint is None
+        assert auth_mechanism.symmetric_key is None
+
+    @pytest.mark.it(
+        "calls method from service operations with device id, module id and previously constructed module"
+    )
+    def test_calls_create_or_update_device_for_certificate_authority(
+        self, mock_module_constructor, mock_service_operations, iothub_registry_manager
+    ):
+        iothub_registry_manager.update_module_with_certificate_authority(
+            device_id=fake_device_id,
+            module_id=fake_module_id,
+            etag=fake_etag,
+            status=fake_status,
+            managed_by=fake_managed_by,
+        )
+
+        assert mock_service_operations.create_or_update_module.call_count == 1
+        assert mock_service_operations.create_or_update_module.call_args[0][0] == fake_device_id
+        assert mock_service_operations.create_or_update_module.call_args[0][1] == fake_module_id
+        assert (
+            mock_service_operations.create_or_update_module.call_args[0][2]
+            == mock_module_constructor.return_value
+        )
 
 
 @pytest.mark.describe("IoTHubRegistryManager -- .get_module()")
